+++ include/H2DE/H2DE.h
@@ -7,7 +7,7 @@
  * Hydren 2 Dimensional Engine (H2DE)
  * https://github.com/Hydrened/H2DE
  * 
- * Version: 3.2.13
+ * Version: 3.2.14
  * 
  * Copyright (C) 2024-2025
  *