+++ include/H2DE/objects/H2DE_text_object.h
@@ -90,8 +90,9 @@ public:
      * @param duration The duration of the animation.
      * @param easing The easing function for the animation.
      * @param pauseSensitive Whether the animation should pause when the game is paused.
+     * @return The ID of the created timeline.
      */
-    friend void H2DE_SetTextObjectFontSize(H2DE_TextObject* textObject, const H2DE_LevelSize& fontSize, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
+    friend unsigned int H2DE_SetTextObjectFontSize(H2DE_TextObject* textObject, const H2DE_LevelSize& fontSize, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
     /**
      * @brief Sets the spacing between characters in the text object.
      * 
@@ -113,8 +114,9 @@ public:
      * @param duration The duration of the animation.
      * @param easing The easing function for the animation.
      * @param pauseSensitive Whether the animation should pause when the game is paused.
+     * @return The ID of the created timeline.
      */
-    friend void H2DE_SetTextObjectSpacing(H2DE_TextObject* textObject, const H2DE_LevelSize& spacing, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
+    friend unsigned int H2DE_SetTextObjectSpacing(H2DE_TextObject* textObject, const H2DE_LevelSize& spacing, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
     /**
      * @brief Sets the text alignment for the text object.
      * 
@@ -145,8 +147,9 @@ public:
      * @param duration The duration of the animation.
      * @param easing The easing function for the animation.
      * @param pauseSensitive Whether the animation should pause when the game is paused.
+     * @return The ID of the created timeline.
      */
-    friend void H2DE_SetTextObjectColor(H2DE_TextObject* textObject, const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
+    friend unsigned int H2DE_SetTextObjectColor(H2DE_TextObject* textObject, const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, bool pauseSensitive);
 
     /**
      * @brief Sets the position of the object.